[Kiet] create + update design + delete calendar

+++ app/Http/Controllers/EventController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Event;
+use Illuminate\Support\Facades\Auth;
 
 class EventController extends Controller
 {
@@ -13,23 +14,47 @@ class EventController extends Controller
         return view('calendar.index');
     }
 
+    public function getEvents()
+    {
+        $userId = Auth::user()->id;
+
+        $events = Event::where('user_id', $userId)->get();
+
+        $formattedEvents = $events->map(function ($event) {
+            return [
+                'id' => $event->id,
+                'title' => $event->title,
+                'start' => $event->start,
+                'end' => $event->end,
+                'description' => $event->description,
+            ];
+        });
+
+        return response()->json($formattedEvents);
+    }
+
     public function store(Request $request)
     {
         $event = Event::create($request->all());
-        return response()->json($event);
+        return redirect()->route('calendar.index')->with('success_create', __('messages.Create_success'));
     }
 
     public function update(Request $request, $id)
     {
         $event = Event::find($id);
         $event->update($request->all());
-        return response()->json($event);
+        return redirect()->route('calendar.index')->with('success_create', __('messages.Create_success'));
     }
 
     public function destroy($id)
     {
         $event = Event::find($id);
-        $event->delete();
-        return response()->json('Event deleted successfully');
+
+        if ($event) {
+            $event->delete();
+            return redirect()->route('calendar.index')->with('success_create', __('messages.Create_success'));
+        }
+    
+        return redirect()->route('calendar.index')->with('success_create', __('messages.Create_success'));
     }
 }
